# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta es una API para subir, transformar y consultar imágenes.",
        "title": "API de Procesamiento de Imágenes",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/images/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Devuelve la metadata y URL de una imagen del usuario autenticado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Obtener información de una imagen",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la imagen",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ImageDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/images/{id}/transform": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Aplica transformaciones (resize, crop, rotación, filtros) a una imagen previamente cargada por el usuario",
                "produces": [
                    "image/png",
                    "image/jpeg",
                    "image/gif"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Aplica transformaciones a una imagen",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la imagen",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Parámetros de transformación",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TransformationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Imagen transformada",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/images/{rest}": {
            "get": {
                "description": "Devuelve el archivo de imagen almacenado en el servidor desde la carpeta local \"uploads\", sin exponer su ruta real.",
                "produces": [
                    "image/jpeg",
                    "image/png",
                    "image/gif"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Servir imagen",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ruta relativa de la imagen (por ejemplo: user123/imagen.jpg)",
                        "name": "rest",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "403": {
                        "description": "Acceso prohibido",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Archivo no encontrado",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Autentica al usuario y devuelve un token JWT si las credenciales son correctas.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Inicia sesión",
                "parameters": [
                    {
                        "description": "Credenciales de usuario",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Crea un nuevo usuario en la base de datos con nombre de usuario y contraseña.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Registra un nuevo usuario",
                "parameters": [
                    {
                        "description": "Datos del usuario a registrar",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permite a un usuario autenticado subir una imagen. Se guarda en MinIO y se almacena la metadata.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Sube una imagen",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Imagen a subir (jpg, jpeg, png, gif)",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.UploadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user-images": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Devuelve las imágenes subidas por el usuario autenticado con soporte de paginación.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Lista imágenes del usuario autenticado",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Número de página",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Cantidad por página",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PaginatedImagesResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Descripción del error"
                }
            }
        },
        "dto.ImageDetailResponse": {
            "type": "object",
            "properties": {
                "format": {
                    "type": "string",
                    "example": "jpeg"
                },
                "height": {
                    "type": "integer",
                    "example": 1080
                },
                "name": {
                    "type": "string",
                    "example": "imagen123.jpg"
                },
                "size": {
                    "type": "integer",
                    "example": 204800
                },
                "url": {
                    "type": "string",
                    "example": "http://localhost:8080/images/rodrick/imagen123.jpg"
                },
                "width": {
                    "type": "integer",
                    "example": 1920
                }
            }
        },
        "dto.ImageItem": {
            "type": "object",
            "properties": {
                "format": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "dto.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "dto.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "dto.PaginatedImagesResponse": {
            "type": "object",
            "properties": {
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ImageItem"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.RegisterRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "dto.TransformationRequest": {
            "type": "object",
            "properties": {
                "transformations": {
                    "type": "object",
                    "properties": {
                        "crop": {
                            "type": "object",
                            "properties": {
                                "height": {
                                    "type": "integer"
                                },
                                "width": {
                                    "type": "integer"
                                },
                                "x": {
                                    "type": "integer"
                                },
                                "y": {
                                    "type": "integer"
                                }
                            }
                        },
                        "filters": {
                            "type": "object",
                            "properties": {
                                "grayscale": {
                                    "type": "boolean"
                                },
                                "sepia": {
                                    "type": "boolean"
                                }
                            }
                        },
                        "format": {
                            "type": "string"
                        },
                        "resize": {
                            "type": "object",
                            "properties": {
                                "height": {
                                    "type": "integer"
                                },
                                "width": {
                                    "type": "integer"
                                }
                            }
                        },
                        "rotate": {
                            "type": "number"
                        }
                    }
                }
            }
        },
        "dto.UploadResponse": {
            "type": "object",
            "properties": {
                "image": {
                    "$ref": "#/definitions/dto.UploadedImageDetail"
                },
                "message": {
                    "type": "string",
                    "example": "Imagen subida exitosamente"
                }
            }
        },
        "dto.UploadedImageDetail": {
            "type": "object",
            "properties": {
                "format": {
                    "type": "string",
                    "example": "jpeg"
                },
                "height": {
                    "type": "integer",
                    "example": 1080
                },
                "name": {
                    "type": "string",
                    "example": "imagen123.jpg"
                },
                "size": {
                    "type": "integer",
                    "example": 204800
                },
                "url": {
                    "type": "string",
                    "example": "http://localhost:8080/images/username/imagen123.jpg"
                },
                "width": {
                    "type": "integer",
                    "example": 1920
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Escribe \"Bearer \" seguido del token JWT obtenido en /login",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}